Reorganize scripts for code reuse and cleaner file structure
