+++ scripts/deploy-contracts.js
@@ -1,0 +1,19 @@
+const { spawn } = require('child_process')
+
+const deployContracts = () => {
+  return new Promise((resolve, reject) => {
+    const truffleMigrate = spawn('../node_modules/.bin/truffle', ['migrate', '--reset', '--compile-all'], { cwd: './contracts' })
+    truffleMigrate.stdout.pipe(process.stdout)
+    truffleMigrate.stderr.pipe(process.stderr)
+
+    truffleMigrate.on('exit', code => {
+      if (code !== 0) {
+        return reject()
+      }
+      console.log('Truffle migrate finished OK.')
+      resolve()
+    })
+  })
+}
+
+module.exports = deployContracts
